refactor: deleted process and upload files from ingest and created a centralized route for both
